refactor: Data Processing Services
### Desc
- 데이터를 추출해내거나, 요약 및 분석을 수행하는 등 데이터 프로세싱을 수행하는 서비스들을 패키지로 관리하자
- `analysis/ai`
  - 서드파티 AI API를 사용하여 데이터 추출을 진행하거나, 데이터 요약을 수행하는 서비스의 모음
  - `OpenAiWeeklyStatsSummaryService`, `OpenAiHomeSummaryService` -> `AiSummaryService`
    - 상기 서비스는 제공되는 정형 데이터의 형식만 다를 뿐, 기능적으로는 큰 차이가 없어 통합하였음. 기술에 의존하지 않는 형식이 좋을 것 같아 이름도 수정
- `analysis`
  - 비정형 전화 데이터를 AI가 가공하여 정형 데이터로 제공해 주는데, 이를 저장하거나 분류하는 등의 데이터 프로세싱을 담당하는 서비스의 모음
  - 앞선 PR에서 `carecall`패키지에 포함하였던 `CareCallDataProcessingService`의 경우에도 이쪽 패키지에 포함되는게 적절해보여 이동하였음
    - `carecall` 패키지쪽은 데이터 처리보다는 케어콜을 수행하는 것에 집중하는 서비스들을 관리하는 패키지로 설정하자.
