+++ src/test/java/com/example/medicare_call/controller/view/MealRecordControllerTest.java
@@ -2,7 +2,7 @@ package com.example.medicare_call.controller.view;
 
 import com.example.medicare_call.dto.DailyMealResponse;
 import com.example.medicare_call.global.jwt.JwtProvider;
-import com.example.medicare_call.service.MealRecordService;
+import com.example.medicare_call.service.MentalAnalysisService;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -31,7 +31,7 @@ class MealRecordControllerTest {
     private MockMvc mockMvc;
 
     @MockBean
-    private MealRecordService mealRecordService;
+    private MentalAnalysisService.MealRecordService mealRecordService;
 
     @MockBean
     private JwtProvider jwtProvider;
